requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.22
langchain-ollama==1.0.1
langchain-qdrant==1.1.0
langsmith==0.7.22
numpy==2.4.3
ollama==0.6.1
orjson==3.11.7
packaging==26.0
portalocker==3.2.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
qdrant-client==1.17.1
requests==2.33.0
requests-toolbelt==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
